+++ zerver/tests/test_reactions.py
@@ -109,11 +109,6 @@ class ReactionEmojiTest(ZulipTestCase):
                 "emoji_name": emoji,
                 "emoji_code": emoji_code,
                 "reaction_type": "unicode_emoji",
-                "user": {
-                    "email": f"user{sender.id}@zulip.testserver",
-                    "id": sender.id,
-                    "full_name": sender.full_name,
-                },
                 "user_id": sender.id,
             }
             # It's important that we preserve the loop order in this
@@ -446,7 +441,6 @@ class ReactionEventTest(ZulipTestCase):
         event_user_ids = set(events[0]["users"])
 
         self.assertEqual(expected_recipient_ids, event_user_ids)
-        self.assertEqual(event["user"]["email"], reaction_sender.email)
         self.assertEqual(event["type"], "reaction")
         self.assertEqual(event["op"], "add")
         self.assertEqual(event["emoji_name"], "smile")
@@ -492,7 +486,6 @@ class ReactionEventTest(ZulipTestCase):
         event_user_ids = set(events[0]["users"])
 
         self.assertEqual(expected_recipient_ids, event_user_ids)
-        self.assertEqual(event["user"]["email"], reaction_sender.email)
         self.assertEqual(event["type"], "reaction")
         self.assertEqual(event["op"], "remove")
         self.assertEqual(event["emoji_name"], "smile")
@@ -1067,9 +1060,6 @@ class ReactionAPIEventTest(EmojiReactionBase):
         event_user_ids = set(events[0]["users"])
 
         self.assertEqual(expected_recipient_ids, event_user_ids)
-        self.assertEqual(event["user"]["user_id"], reaction_sender.id)
-        self.assertEqual(event["user"]["email"], reaction_sender.email)
-        self.assertEqual(event["user"]["full_name"], reaction_sender.full_name)
         self.assertEqual(event["type"], "reaction")
         self.assertEqual(event["op"], "add")
         self.assertEqual(event["message_id"], pm_id)
@@ -1112,9 +1102,6 @@ class ReactionAPIEventTest(EmojiReactionBase):
         event_user_ids = set(events[0]["users"])
 
         self.assertEqual(expected_recipient_ids, event_user_ids)
-        self.assertEqual(event["user"]["user_id"], reaction_sender.id)
-        self.assertEqual(event["user"]["email"], reaction_sender.email)
-        self.assertEqual(event["user"]["full_name"], reaction_sender.full_name)
         self.assertEqual(event["type"], "reaction")
         self.assertEqual(event["op"], "remove")
         self.assertEqual(event["message_id"], pm_id)